Fixes Issue [#59](https://github.com/chef-cookbooks/chef-splunk/issues/59)   * converts `splunk_app` resource to modern style
