+++ resources/splunk_installer.rb
@@ -1,5 +1,6 @@
 #
 # Author: Dang H. Nguyen <dang.nguyen@disney.com>
+# Copyright:: 2019-2020
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
